prepare for mgmt by GHA
