+++ .gitignore
@@ -1,14 +1,22 @@
+.idea
+.code
 **/vendor/
-**/node_modules/
 .pulumi
 **/bin/
 **/obj/
 Pulumi.*.yaml
+**/node_modules/
 
 **/command-output/
 
 .idea/
 *.iml
 
-**/yarn.lock
+yarn.lock
 **/pulumiManifest.go
+
+ci-scripts
+**/schema.go
+**/version.txt
+**/nuget
+**/dist
