--- .golangci.yaml
@@ -1,19 +1,0 @@
-linters:
-  enable-all: false
-  enable:
-    - deadcode
-    - errcheck
-    - goconst
-    - gofmt
-    - golint
-    - gosec
-    - govet
-    - ineffassign
-    - interfacer
-    - lll
-    - megacheck
-    - misspell
-    - nakedret
-    - structcheck
-    - unconvert
-    - varcheck
