--- provider/cmd/pulumi-resource-f5bigip/.gitignore
@@ -1,1 +1,0 @@
-schema.go
